detection for devices and device specific scaling (still needs work). ios/ipados WIDGETS

+++ packages/flutter_drawing_board/lib/src/paint_extension/ex_color.dart
@@ -19,8 +19,8 @@ Color? stringToColor(String data) {
 }
 
 int colorToInt(Color color) {
-  return (color.alpha * 255).toInt() << 24 |
-      (color.red * 255).toInt() << 16 |
-      (color.green * 255).toInt() << 8 |
-      (color.blue * 255).toInt();
+  return (color.alpha * 255) << 24 |
+      (color.red * 255) << 16 |
+      (color.green * 255) << 8 |
+      (color.blue * 255);
 }
